# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Swagger for Jobbi Back-end\n\nSchemes: [http, https]",
    "title": "Jobbi Back-end",
    "contact": {
      "name": "Maxime Denuit",
      "email": "maxime@nightborn.be"
    },
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
...
    "/searches/{searchId}": {
      "get": {
        "description": "type id struct\nGet search by id.\nReturn search",
        "produces": [
          "application/json"
        ],
        "tags": [
          "searches"
        ],
        "operationId": "GetSearchById",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SearchDTO"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "description": "type id struct\nCreate search.\nReturn search",
        "produces": [
          "application/json"
        ],
        "tags": [
          "searches"
        ],
        "operationId": "ModifySearch",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          },
          {
            "description": "search",
            "name": "search",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/SearchDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SearchDTO"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/searches/{searchId}/offers": {
      "get": {
        "description": "type id struct\nGet offers by searchId.\nReturn offer",
        "produces": [
          "application/json"
        ],
        "tags": [
          "offers"
        ],
        "operationId": "GetOffersBySearchId",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/OfferDTO"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "description": "type id struct\nCreate offer.\nReturn offer",
        "produces": [
          "application/json"
        ],
        "tags": [
          "offers"
        ],
        "operationId": "AddOffer",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          },
          {
            "description": "offer",
            "name": "offer",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/OfferDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/OfferDTO"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/searches/{searchId}/offers/{offerId}": {
      "get": {
        "description": "type id struct\nGet offer by id.\nReturn offer",
        "produces": [
          "application/json"
        ],
        "tags": [
          "offers"
        ],
        "operationId": "GetOfferById",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "test",
            "name": "offerId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/OfferDTO"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "description": "type id struct\nModify offer.\nReturn offer",
        "produces": [
          "application/json"
        ],
        "tags": [
          "offers"
        ],
        "operationId": "ModifyOffer",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "test",
            "name": "offerId",
            "in": "path",
            "required": true
          },
          {
            "description": "offer",
            "name": "offer",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/OfferDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/OfferDTO"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "description": "type id struct\nDelete offer.\nReturn true or error",
        "produces": [
          "application/json"
        ],
        "tags": [
          "offers"
        ],
        "operationId": "DeleteOffer",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "test",
            "name": "offerId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "definitions": {
...
    "OfferDTO": {
      "description": "An offer",
      "type": "object",
      "properties": {
        "description": {
          "description": "The description",
          "type": "string",
          "x-go-name": "Description"
        },
        "id": {
          "description": "The id",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Id"
        },
        "link": {
          "description": "The link",
          "type": "string",
          "x-go-name": "Link"
        },
        "searchId": {
          "description": "the search id",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "SearchID"
        },
        "title": {
          "description": "The title",
          "type": "string",
          "x-go-name": "Title"
        },
        "userId": {
          "description": "the user id",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "github.com/philvc/jobbi-api/contract"
    },
...
    "SearchDTO": {
      "description": "An search",
      "type": "object",
      "required": [
        "id",
        "description",
        "title"
      ],
      "properties": {
        "description": {
          "description": "The description",
          "type": "string",
          "x-go-name": "Description"
        },
        "id": {
          "description": "The id",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Id"
        },
        "title": {
          "description": "The title",
          "type": "string",
          "x-go-name": "Title"
        },
        "userId": {
          "description": "UserId",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "github.com/philvc/jobbi-api/contract"
    },
...
  }
}